Move model class to model package
