+++ app/src/main/java/com/github/mik629/android/fundamentals/data/network/model/MovieItem.kt
@@ -1,4 +1,4 @@
-package com.github.mik629.android.fundamentals.data.network
+package com.github.mik629.android.fundamentals.data.network.model
 
 data class MovieItem(
     val title: String,
